+++ src/expressions/expression.rs
@@ -58,17 +58,12 @@ impl Expressionable for Source {
     fn parse_unary_atom(input: &mut ParseStreamStack<Self>) -> ParseResult<UnaryAtom<Self>> {
         Ok(match input.peek_grammar() {
             SourcePeekMatch::Command(_) => UnaryAtom::Leaf(Self::Leaf::Command(input.parse()?)),
-            SourcePeekMatch::Variable(Grouping::Grouped) => {
+            SourcePeekMatch::EmbeddedVariable => {
                 return input.parse_err(
                     "In an expression, the # variable prefix is not allowed. The # prefix should only be used when embedding a variable into an output stream.",
                 )
             }
-            SourcePeekMatch::Variable(Grouping::Flattened) => {
-                return input.parse_err(
-                    "In an expression, the #.. variable prefix is not allowed. The # prefix should only be used when embedding a variable into an output sream.",
-                )
-            }
-            SourcePeekMatch::EmbeddedExpression(_) => {
+            SourcePeekMatch::EmbeddedExpression => {
                 UnaryAtom::Leaf(Self::Leaf::EmbeddedExpression(input.parse()?))
             }
             SourcePeekMatch::ExplicitTransformStream | SourcePeekMatch::Transformer(_) => {