+++ src/interpretation/variable.rs
@@ -28,11 +28,9 @@ pub(crate) trait IsVariable: HasSpanRange {
         grouping: Grouping,
         output: &mut OutputStream,
     ) -> ExecutionResult<()> {
-        self.binding(interpreter)?.into_shared()?.output_to(
-            grouping,
-            output,
-            StreamOutputBehaviour::Standard,
-        )
+        self.binding(interpreter)?
+            .into_shared()?
+            .output_to(grouping, output)
     }
 
     fn binding(&self, interpreter: &Interpreter) -> ExecutionResult<VariableBinding> {
@@ -43,65 +41,12 @@ pub(crate) trait IsVariable: HasSpanRange {
 }
 
 #[derive(Clone)]
-pub(crate) enum MarkedVariable {
-    Grouped(GroupedVariable),
-    Flattened(FlattenedVariable),
-}
-
-impl Parse<Source> for MarkedVariable {
-    fn parse(input: ParseStream<Source>) -> ParseResult<Self> {
-        if input.peek2(Token![..]) {
-            Ok(MarkedVariable::Flattened(input.parse()?))
-        } else {
-            Ok(MarkedVariable::Grouped(input.parse()?))
-        }
-    }
-}
-
-impl HasSpanRange for MarkedVariable {
-    fn span_range(&self) -> SpanRange {
-        match self {
-            MarkedVariable::Grouped(variable) => variable.span_range(),
-            MarkedVariable::Flattened(variable) => variable.span_range(),
-        }
-    }
-}
-
-impl HasSpanRange for &MarkedVariable {
-    fn span_range(&self) -> SpanRange {
-        <MarkedVariable as HasSpanRange>::span_range(self)
-    }
-}
-
-impl IsVariable for MarkedVariable {
-    fn get_name(&self) -> String {
-        match self {
-            MarkedVariable::Grouped(variable) => variable.get_name(),
-            MarkedVariable::Flattened(variable) => variable.get_name(),
-        }
-    }
-}
-
-impl Interpret for &MarkedVariable {
-    fn interpret_into(
-        self,
-        interpreter: &mut Interpreter,
-        output: &mut OutputStream,
-    ) -> ExecutionResult<()> {
-        match self {
-            MarkedVariable::Grouped(variable) => variable.interpret_into(interpreter, output),
-            MarkedVariable::Flattened(variable) => variable.interpret_into(interpreter, output),
-        }
-    }
-}
-
-#[derive(Clone)]
-pub(crate) struct GroupedVariable {
+pub(crate) struct EmbeddedVariable {
     marker: Token![#],
     variable_name: Ident,
 }
 
-impl Parse<Source> for GroupedVariable {
+impl Parse<Source> for EmbeddedVariable {
     fn parse(input: ParseStream<Source>) -> ParseResult<Self> {
         input.try_parse_or_error(
             |input| {
@@ -115,23 +60,23 @@ impl Parse<Source> for GroupedVariable {
     }
 }
 
-impl IsVariable for GroupedVariable {
+impl IsVariable for EmbeddedVariable {
     fn get_name(&self) -> String {
         self.variable_name.to_string()
     }
 }
 
-impl Interpret for &GroupedVariable {
+impl Interpret for &EmbeddedVariable {
     fn interpret_into(
         self,
         interpreter: &mut Interpreter,
         output: &mut OutputStream,
     ) -> ExecutionResult<()> {
-        self.substitute_into(interpreter, Grouping::Grouped, output)
+        self.substitute_into(interpreter, Grouping::Flattened, output)
     }
 }
 
-impl InterpretToValue for &GroupedVariable {
+impl InterpretToValue for &EmbeddedVariable {
     type OutputValue = ExpressionValue;
 
     fn interpret_to_value(
@@ -142,78 +87,21 @@ impl InterpretToValue for &GroupedVariable {
     }
 }
 
-impl HasSpanRange for GroupedVariable {
+impl HasSpanRange for EmbeddedVariable {
     fn span_range(&self) -> SpanRange {
         SpanRange::new_between(self.marker.span, self.variable_name.span())
     }
 }
 
-impl HasSpanRange for &GroupedVariable {
+impl HasSpanRange for &EmbeddedVariable {
     fn span_range(&self) -> SpanRange {
-        <GroupedVariable as HasSpanRange>::span_range(self)
+        <EmbeddedVariable as HasSpanRange>::span_range(self)
     }
 }
 
-impl core::fmt::Display for GroupedVariable {
+impl core::fmt::Display for EmbeddedVariable {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         write!(f, "#{}", self.variable_name)
-    }
-}
-
-#[derive(Clone)]
-pub(crate) struct FlattenedVariable {
-    marker: Token![#],
-    #[allow(unused)]
-    flatten: Token![..],
-    variable_name: Ident,
-}
-
-impl Parse<Source> for FlattenedVariable {
-    fn parse(input: ParseStream<Source>) -> ParseResult<Self> {
-        input.try_parse_or_error(
-            |input| {
-                Ok(Self {
-                    marker: input.parse()?,
-                    flatten: input.parse()?,
-                    variable_name: input.parse_any_ident()?,
-                })
-            },
-            "Expected #..variable",
-        )
-    }
-}
-
-impl IsVariable for FlattenedVariable {
-    fn get_name(&self) -> String {
-        self.variable_name.to_string()
-    }
-}
-
-impl Interpret for &FlattenedVariable {
-    fn interpret_into(
-        self,
-        interpreter: &mut Interpreter,
-        output: &mut OutputStream,
-    ) -> ExecutionResult<()> {
-        self.substitute_into(interpreter, Grouping::Flattened, output)
-    }
-}
-
-impl HasSpanRange for FlattenedVariable {
-    fn span_range(&self) -> SpanRange {
-        SpanRange::new_between(self.marker.span, self.variable_name.span())
-    }
-}
-
-impl HasSpanRange for &FlattenedVariable {
-    fn span_range(&self) -> SpanRange {
-        <FlattenedVariable as HasSpanRange>::span_range(self)
-    }
-}
-
-impl core::fmt::Display for FlattenedVariable {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        write!(f, "#..{}", self.variable_name)
     }
 }