+++ src/lib.rs
@@ -568,9 +568,7 @@ fn preinterpret_run_internal(input: TokenStream) -> SynResult<TokenStream> {
 
     let interpreted_stream = block_content
         .evaluate(&mut interpreter, Span::call_site().into())
-        .and_then(|x| {
-            x.into_new_output_stream(Grouping::Flattened, StreamOutputBehaviour::PermitArrays)
-        })
+        .and_then(|x| x.into_new_output_stream(Grouping::Flattened))
         .convert_to_final_result()?;
 
     unsafe {
@@ -628,12 +626,7 @@ mod benchmarking {
             let mut interpreter = Interpreter::new();
             block_content
                 .evaluate(&mut interpreter, Span::call_site().into())
-                .and_then(|x| {
-                    x.into_new_output_stream(
-                        Grouping::Flattened,
-                        StreamOutputBehaviour::PermitArrays,
-                    )
-                })
+                .and_then(|x| x.into_new_output_stream(Grouping::Flattened))
                 .convert_to_final_result()
         });
         let interpreted_stream = interpreted_stream?;