+++ tests/compilation_failures/core/extend_flattened_variable.rs
@@ -1,8 +1,0 @@
-use preinterpret::*;
-
-fn main() {
-    stream! {
-        [!set! #variable = 1]
-        [!set! #..variable += 2]
-    }
-}