+++ tests/compilation_failures/core/set_flattened_variable.rs
@@ -1,7 +1,0 @@
-use preinterpret::*;
-
-fn main() {
-    stream! {
-        [!set! #..variable = 1]
-    }
-}