+++ tests/compilation_failures/expressions/embedded_variable_in_expression.rs
@@ -1,0 +1,8 @@
+use preinterpret::*;
+
+fn main() {
+    let _ = run! {
+        let partial_sum = [!stream! + 2];
+        5 #partial_sum
+    };
+}