+++ tests/compilation_failures/expressions/flattened_variables_in_expressions.rs
@@ -1,7 +1,0 @@
-use preinterpret::*;
-
-fn main() {
-    let _ = stream! {
-        #(partial_sum = [!stream! + 2]; 5 #..partial_sum)
-    };
-}