+++ tests/compilation_failures/tokens/intersperse_stream_input_variable_issue.rs
@@ -2,9 +2,9 @@ use preinterpret::*;
 
 fn main() {
     stream! {
-        [!set! #x = 1 2]
+        #(let x = %[1 2];)
         [!intersperse! %{
-            items: #..x,
+            items: #x,
             separator: []
         }]
     }