+++ tests/compilation_failures/transforming/destructure_with_punct_flattened_variable.rs
@@ -1,5 +1,0 @@
-use preinterpret::*;
-
-fn main() {
-    stream!([!let! @(#..x = @PUNCT) = @]);
-}